From Java Advanced Exams
You are asked to cook pastries and cakes for the near pastry shop, because your recipes are so great. You need to mix liquids with ingredients to cook the required delicacies.
First, you will be given a sequence of integers, representing liquids. Afterwards, you will be given another sequence of integers representing ingredients.
You need to start from the first liquid and try to mix it with the last ingredient. If the sum of their values is equal to any of the items in the table below – cook the food corresponding to the value and remove both the liquid and the ingredient. Otherwise, remove only the liquid and increase the value of the ingredient by 3. You need to stop combining when you have no more liquids or ingredients.

Input
•	On the first line, you will receive the integers representing the liquids, separated by a single space. 
•	On the second line, you will receive the integers representing the ingredients, separated by a single space.
Output
•	On the first line of output print one of the following ouputs:
o	"Great! You succeeded in cooking all the food!" -if you have at least
one of each of the foods, after completing combining.
o	"What a pity! You didn't have enough materials to cook everything." – if you did not collect one of each of the foods, after completing combining.
•	On the second line - print all liquids you have left:
o	If there are no liquids: "Liquids left: none"
o	If there are liquids: "Liquids left: {liquid1}, {liquid2}, {liquid3}, (…)"
•	On the third line - print all physical materials you have left:
o	If there are no items: "Ingredients left: none"
o	If there are items: "Ingredients left: {ingredient}, {ingredient}, {ingredient}, (…)"
•	Then, you need to print all Advanced Materials and the amount you have of them, ordered:
o	"Biscuit: {amount}"
o	"Cake: {amount}"
o	"Pie: {amount}"
o	"Pastry: {amount}"