From Java Advanced Exams
First you will receive a sequence of integers, representing the number of ingredients in a single bucket. After that you will be given another sequence of integers - the freshness level of the ingredients.
Your task is to mix them so you can produce the cocktails, listed in the table below with the exact freshness level. 
To mix a cocktail, you have to take the first bucket of ingredient and the last freshness level value. The total freshness level is calculated by their multiplication. If the product of this operation equals one of the levels described in the table, you make the cocktail and remove both bucket with ingredients and freshness value. Otherwise you should remove the freshness level, increase the ingredient value by 5, then remove it from the first position and add it at the end. In case you have an ingredient with value 0 you have to remove it and continue mixing the cocktails.
You need to stop making cocktails when you run out of buckets with ingredients or freshness level values.
Your task is considered done if you make at least four cocktails - one of each type.
Input
•	The first line of input will represent the values of buckets with ingredients - integers, separated by single space.
•	On the second line you will be given the freshness values - integers again, separated by single space.
Output
•	On the first line of output - print whether you've succeeded in preparing the cocktails
o	"It's party time! The cocktails are ready!".
o	"What a pity! You didn't manage to prepare all cocktails.".
•	On the next output line - print the sum of the ingredients only if they are left any 
o	"Ingredients left: {sum of the left ingredients}".
•	On the last few lines you have to print the cocktails you have made at least once, ordered alphabetically in the format:
" # {cocktail name} --> {amount}".