From Database Access with JDBC
Write a program that reads information about a minion and its villain and adds it to the database. In case the town of the minion is not in the database, insert it as well. In case the villain is not present in the database, add him too with default evilness factor of “evil”. Finally, set the new minion to be servant of the villain. Print appropriate messages after each operation – see the examples.